Sparse GPT Algorithm
This is the sparseGPT code based on IST-DASLab project.

I followed the same coding principles as used in the lit-llama gptq code.

I created a file called sparsification  which is the algorithm for SparseGPT and a folder called sparsify/sparsegpt.py to run the algorithm on the the model in the checkpoint_path.

This is my first contribution to the project, If I missed some household admin I apologise in advance.

Key Notes:

    1. The source code of SparseGPT consist of the quantization algorithm similar to GPTQ, however I removed this code because we already have GPTQ in the lit-llama source code.

2. I'm still in the waiting list for the Llama weights 7B.