Add domain specific evaluation of predictions
Add `evaluation.py`, which performs an evaluation of a (fine-tuned)
model based on our domain specific expected behavior. For this,
treat a prediction as a kind of multiclass multioutput classification
of its chat history. A prediction is parsed into a set of calls, each
having a sequence of arguments.

Since `evaluation.py` is very different from the perplexity based
evaluations in the `evaluate` folder, and expects a dataset that can be
created with other scripts in this folder, put `evaluation.py` in this
folder, too.

Add `README.md` to describe this prediction-evaluation-pipeline.

TODO:
1) add cosine similarity comparisons where similar text
(e.g. the msg argument) is sufficient.
2) construct a full confusion matrix and report it to tensorboard.